refactor lecturer-interface
