+++ src/components/CourseComponents/CloneCourseModal.vue
@@ -2,10 +2,10 @@
 import { defineEmits } from "vue/dist/vue";
 import { ref } from "vue";
 import { useToast } from "vue-toastification";
-import { ICourse } from "@/ts/models";
+import { ICourse } from "@/ts/modals/models";
 import { defineProps, watch } from "vue";
-import { postCloneCourse } from "@/ts/course-rest-client";
-import { validateSemester } from "@/ts/validate";
+import { postCloneCourse } from "@/ts/rest-clients/course-rest-client";
+import { validateSemester } from "@/ts/validation/validate";
 
 const props = defineProps<{
   showModal: boolean;
