+++ src/components/EditNPCModal.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import { defineProps, defineEmits, ref, watch } from "vue";
-import { INPC } from "@/ts/models";
+import { INPC } from "@/ts/modals/models";
 
 const props = defineProps<{
   npc: INPC;
