+++ src/ts/rest-clients/finitequiz-rest-client.ts
@@ -2,7 +2,7 @@ import axios, { AxiosResponse } from "axios";
 
 import config from "@/config";
 
-import { IFinitequizConfiguration } from "@/ts/models";
+import { IFinitequizConfiguration } from "@/ts/modals/models";
 
 export async function postFinitequizConfig(
   finitequizConfig: IFinitequizConfiguration
