+++ tests/unit/NPCsView.spec.ts
@@ -2,7 +2,7 @@ import { mount, VueWrapper, flushPromises } from "@vue/test-utils";
 import mockAxios from "jest-mock-axios";
 import router from "@/router/index";
 import NPCView from "@/views/NPCsView.vue";
-import { Dungeon, INPC, NPC, World } from "@/ts/models";
+import { Dungeon, INPC, NPC, World } from "@/ts/modals/models";
 import BootstrapVue3 from "bootstrap-vue-3";
 import config from "@/config";
